@@ -1,6 +1,6 @@
 {
   "name": "renren-fast-vue",
-  "version": "1.0.0",
+  "version": "1.2.1",
   "description": "renren-fast-vue基于vue、element-ui构建开发，实现renren-fast后台管理前端功能，提供一套更优的前端解决方案。",
   "author": "daxiong.yang <daxiong.yang@qq.com>",
   "private": true,
@@ -17,12 +17,16 @@
     "axios": "0.17.1",
     "babel-plugin-component": "0.10.1",
     "babel-polyfill": "6.26.0",
-    "element-ui": "2.1.0",
-    "gulp": "^3.9.1",
-    "gulp-concat": "^2.6.1",
-    "gulp-load-plugins": "^1.5.0",
-    "gulp-replace": "^0.6.1",
-    "gulp-shell": "^0.6.5",
+    "element-ui": "2.3.2",
+    "gulp": "3.9.1",
+    "gulp-concat": "2.6.1",
+    "gulp-load-plugins": "1.5.0",
+    "gulp-replace": "0.6.1",
+    "gulp-shell": "0.6.5",
+    "lodash": "4.17.5",
+    "node-sass": "4.9.0",
+    "sass-loader": "6.0.6",
+    "svg-sprite-loader": "3.7.3",
     "vue": "2.5.2",
     "vue-cookie": "1.1.4",
     "vue-router": "3.0.1",
@@ -52,7 +56,7 @@
     "eslint-plugin-html": "3.0.0",
     "eslint-plugin-import": "2.7.0",
     "eslint-plugin-node": "5.2.0",
-    "eslint-plugin-promise": "3.4.0",
+    "eslint-plugin-promise": "3.5.0",
     "eslint-plugin-standard": "3.0.1",
     "eventsource-polyfill": "0.9.6",
     "extract-text-webpack-plugin": "3.0.0",
@@ -63,14 +67,12 @@
     "jest-serializer-vue": "0.3.0",
     "nightwatch": "0.9.12",
     "node-notifier": "5.1.2",
-    "node-sass": "4.7.2",
     "optimize-css-assets-webpack-plugin": "3.2.0",
     "ora": "1.2.0",
     "portfinder": "1.0.13",
     "postcss-import": "11.0.0",
     "postcss-loader": "2.0.8",
     "rimraf": "2.6.0",
-    "sass-loader": "6.0.6",
     "selenium-server": "3.0.1",
     "semver": "5.3.0",
     "shelljs": "0.7.6",
@@ -86,8 +88,8 @@
     "webpack-merge": "4.1.0"
   },
   "engines": {
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 8.11.1",
+    "npm": ">= 5.6.0"
   },
   "browserslist": [
     "> 1%",
