@@ -1,54 +1,153 @@
+/**
+ * 全站路由配置
+ *
+ * 建议:
+ * 1. 代码中路由统一使用name属性跳转(不使用path属性)
+ */
 import Vue from 'vue'
 import Router from 'vue-router'
-
-// 开发环境不使用懒加载, 因为懒加载页面太多的话会造成webpack热更新太慢, 所以只有开发环境使用懒加载
-const _import = require('./import-' + process.env.NODE_ENV)
+import http from '@/utils/httpRequest'
+import { isURL } from '@/utils/validate'
 
 Vue.use(Router)
 
-// 路由定义使用说明:
-// 1. 代码中路由统一使用name属性跳转.
-// 2. 开放path属性用做简短路由, 比如: '/a1', 访问地址: www.renren.io/#/a1
-export default new Router({
+// 开发环境不使用懒加载, 因为懒加载页面太多的话会造成webpack热更新太慢, 所以只有生产环境使用懒加载
+const _import = require('./import-' + process.env.NODE_ENV)
+
+// 全局路由(无需嵌套上左右整体布局)
+const globalRoutes = [
+  { path: '/404', component: _import('common/404'), name: '404', meta: { title: '404未找到' } },
+  { path: '/login', component: _import('common/login'), name: 'login', meta: { title: '登录' } }
+]
+
+// 主入口路由(需嵌套上左右整体布局)
+const mainRoutes = {
+  path: '/',
+  component: _import('main'),
+  name: 'main',
+  redirect: { name: 'home' },
+  meta: { title: '主入口整体布局' },
+  children: [
+    // 通过meta对象设置路由展示方式
+    // 1. isTab: 是否通过tab展示内容, true: 是, false: 否
+    // 2. iframeUrl: 是否通过iframe嵌套展示内容, '以http[s]://开头': 是, '': 否
+    // 提示: 如需要通过iframe嵌套展示内容, 但不通过tab打开, 请自行创建组件使用iframe处理!
+    { path: '/home', component: _import('common/home'), name: 'home', meta: { title: '首页' } },
+    { path: '/theme', component: _import('common/theme'), name: 'theme', meta: { title: '主题' } },
+    { path: '/demo-echarts', component: _import('demo/echarts'), name: 'demo-echarts', meta: { title: 'demo-echarts', isTab: true } },
+    { path: '/demo-ueditor', component: _import('demo/ueditor'), name: 'demo-ueditor', meta: { title: 'demo-ueditor', isTab: true } }
+  ],
+  beforeEnter (to, from, next) {
+    let token = Vue.cookie.get('token')
+    if (!token || !/\S/.test(token)) {
+      next({ name: 'login' })
+    }
+    next()
+  }
+}
+
+const router = new Router({
   mode: 'hash',
-  routes: [
-    { path: '/404', component: _import('error/404'), name: '404', desc: '404未找到' },
-    { path: '/login', component: _import('login/index'), name: 'login', desc: '登录' },
-    {
-      path: '/',
-      component: _import('layout/index'),
-      name: 'layout',
-      redirect: { name: 'home' },
-      desc: '上左右整体布局',
-      children: [
-        { path: '/home', component: _import('home/index'), name: 'home', desc: '首页' },
-        {
-          path: '/content-tabs',
-          component: _import('layout/content-tabs'),
-          name: 'content-tabs',
-          redirect: { name: '404' },
-          desc: '内容需通过tabs展示',
-          children: [
-            // 以'/n'开头统一拦截, 标记为左侧菜单导航tab展示内容方式路由
-            { path: '/n/user', component: _import('user/index'), name: 'user', desc: '管理员管理' },
-            { path: '/n/role', component: _import('role/index'), name: 'role', desc: '角色管理' },
-            { path: '/n/menu', component: _import('menu/index'), name: 'menu', desc: '菜单管理' },
-            { path: '/n/sql', component: _import('sql/index'), name: 'sql', desc: 'SQL监控' },
-            { path: '/n/schedule', component: _import('schedule/index'), name: 'schedule', desc: '定时任务' },
-            { path: '/n/config', component: _import('config/index'), name: 'config', desc: '参数管理' },
-            { path: '/n/oss', component: _import('oss/index'), name: 'oss', desc: '文件上传' },
-            { path: '/n/log', component: _import('log/index'), name: 'log', desc: '系统日志' }
-          ]
-        }
-      ],
-      beforeEnter (to, from, next) {
-        let token = Vue.cookie.get('token')
-        if (!token || !/\S/.test(token)) {
-          next({ name: 'login' })
-        }
+  scrollBehavior: () => ({ y: 0 }),
+  isAddDynamicMenuRoutes: false, // 是否已经添加动态(菜单)路由
+  routes: globalRoutes.concat(mainRoutes)
+})
+
+router.beforeEach((to, from, next) => {
+  // 添加动态(菜单)路由
+  // 1. 已经添加 or 全局路由, 直接访问
+  // 2. 获取菜单列表, 添加并保存本地存储
+  if (router.options.isAddDynamicMenuRoutes || fnCurrentRouteType(to) === 'global') {
+    next()
+  } else {
+    http({
+      url: http.adornUrl('/sys/menu/nav'),
+      method: 'get',
+      params: http.adornParams()
+    }).then(({data}) => {
+      if (data && data.code === 0) {
+        fnAddDynamicMenuRoutes(data.menuList)
+        router.options.isAddDynamicMenuRoutes = true
+        sessionStorage.setItem('menuList', JSON.stringify(data.menuList || '[]'))
+        sessionStorage.setItem('permissions', JSON.stringify(data.permissions || '[]'))
+        next({ ...to, replace: true })
+      } else {
+        sessionStorage.setItem('menuList', '[]')
+        sessionStorage.setItem('permissions', '[]')
         next()
       }
-    },
-    { path: '*', redirect: { name: '404' } }
-  ]
+    })
+  }
 })
+
+/**
+ * 判断当前路由类型, global: 全局路由, main: 主入口路由
+ * @param {*} route 当前路由
+ */
+function fnCurrentRouteType (route) {
+  var temp = []
+  for (var i = 0; i < globalRoutes.length; i++) {
+    if (route.path === globalRoutes[i].path) {
+      return 'global'
+    } else if (globalRoutes[i].children && globalRoutes[i].children.length >= 1) {
+      temp = temp.concat(globalRoutes[i].children)
+    }
+  }
+  return temp.length >= 1 ? fnCurrentRouteType(route, temp) : 'main'
+}
+
+/**
+ * 添加动态(菜单)路由
+ * @param {*} menuList 菜单列表
+ * @param {*} routes 递归创建的动态(菜单)路由
+ */
+function fnAddDynamicMenuRoutes (menuList = [], routes = []) {
+  var temp = []
+  for (var i = 0; i < menuList.length; i++) {
+    if (menuList[i].list && menuList[i].list.length >= 1) {
+      temp = temp.concat(menuList[i].list)
+    } else if (menuList[i].url && /\S/.test(menuList[i].url)) {
+      menuList[i].url = menuList[i].url.replace(/^\//, '')
+      var route = {
+        path: menuList[i].url.replace('/', '-'),
+        component: null,
+        name: menuList[i].url.replace('/', '-'),
+        meta: {
+          menuId: menuList[i].menuId,
+          title: menuList[i].name,
+          isDynamic: true,
+          isTab: true,
+          iframeUrl: ''
+        }
+      }
+      // url以http[s]://开头, 通过iframe展示
+      if (isURL(menuList[i].url)) {
+        route['path'] = `i-${menuList[i].menuId}`
+        route['name'] = `i-${menuList[i].menuId}`
+        route['meta']['iframeUrl'] = menuList[i].url
+      } else {
+        try {
+          route['component'] = _import(`modules/${menuList[i].url}`) || null
+        } catch (e) {}
+      }
+      routes.push(route)
+    }
+  }
+  if (temp.length >= 1) {
+    fnAddDynamicMenuRoutes(temp, routes)
+  } else {
+    mainRoutes.name = 'main-dynamic'
+    mainRoutes.children = routes
+    router.addRoutes([
+      mainRoutes,
+      { path: '*', redirect: { name: '404' } }
+    ])
+    sessionStorage.setItem('dynamicMenuRoutes', JSON.stringify(mainRoutes.children || '[]'))
+    console.log('\n')
+    console.log('%c!<-------------------- 动态(菜单)路由 s -------------------->', 'color:blue')
+    console.log(mainRoutes.children)
+    console.log('%c!<-------------------- 动态(菜单)路由 e -------------------->', 'color:blue')
+  }
+}
+
+export default router
