@@ -4,16 +4,19 @@ var path = require('path');
 var del  = require('del');
 
 var distPath    = path.resolve('./dist');
-var version     = ''; // 版本号(静态资源文件夹名称)
-var versionPath = ''; // 版本号路径(静态资源文件夹路径)
-var env         = ''; // 环境
+var version     = ''; // 版本号
+var versionPath = ''; // 版本号路径
+var env         = ''; // 运行环境
 
-// 创建版本号 / 版本号路径
+// 创建版本号(年月日时分)
 (function () {
   var d = new Date();
-  version = (d.getFullYear().toString().slice(2))
-    + ((d.getMonth() + 1) >= 10 ? (d.getMonth() + 1) : '0' + (d.getMonth() + 1))
-    + (d.getDate() >= 10 ? d.getDate() : '0' + d.getDate());
+  var yy = d.getFullYear().toString().slice(2);
+  var MM = d.getMonth() + 1 >= 10 ? (d.getMonth() + 1) : '0' + (d.getMonth() + 1);
+  var DD = d.getDate() >= 10 ? d.getDate() : '0' + d.getDate();
+  var h  = d.getHours() >= 10 ? d.getHours() : '0' + d.getHours();
+  var mm = d.getMinutes() >= 10 ? d.getMinutes() : '0' + d.getMinutes();
+  version = yy + MM + DD + h + mm;
   versionPath = distPath + '/' + version;
 })();
 
@@ -21,27 +24,27 @@ var env         = ''; // 环境
 gulp.task('build', $.shell.task([ 'node build/build.js' ]));
 
 // 创建版本号目录
-gulp.task('create:versionPath', ['build'], function () {
+gulp.task('create:versionCatalog', ['build'], function () {
   return gulp.src(`${distPath}/static/**/*`)
     .pipe(gulp.dest(`${versionPath}/static/`))
 });
 
 // 替换${versionPath}/static/js/manifest.js window.SITE_CONFIG.cdnUrl占位变量
-gulp.task('replace:cdnUrl', ['create:versionPath'], function () {
+gulp.task('replace:cdnUrl', ['create:versionCatalog'], function () {
   return gulp.src(`${versionPath}/static/js/manifest.js`)
-    .pipe($.replace(/"\.\/"/g, 'window.SITE_CONFIG.cdnUrl + "/"'))
+    .pipe($.replace(new RegExp(`"${require('./config').build.assetsPublicPath}"`, 'g'), 'window.SITE_CONFIG.cdnUrl + "/"'))
     .pipe(gulp.dest(`${versionPath}/static/js/`))
 });
 
-// 替换${versionPath}/static/config/index-${env}.js window.SITE_CONFIG.staticFileName配置变量
-gulp.task('replace:staticFileName', ['create:versionPath'], function () {
+// 替换${versionPath}/static/config/index-${env}.js window.SITE_CONFIG['version']配置变量
+gulp.task('replace:version', ['create:versionCatalog'], function () {
   return gulp.src(`${versionPath}/static/config/index-${env}.js`)
-    .pipe($.replace(/window.SITE_CONFIG.staticFileName = \'.*\'/g, `window.SITE_CONFIG.staticFileName = \'${version}\'`))
+    .pipe($.replace(/window.SITE_CONFIG\['version'\] = '.*'/g, `window.SITE_CONFIG['version'] = '${version}'`))
     .pipe(gulp.dest(`${versionPath}/static/config/`))
 });
 
 // 合并${versionPath}/static/config/[index-${env}, init].js 至 ${distPath}/config/index.js
-gulp.task('concat:config', ['replace:staticFileName'], function () {
+gulp.task('concat:config', ['replace:version'], function () {
   return gulp.src([`${versionPath}/static/config/index-${env}.js`, `${versionPath}/static/config/init.js`])
     .pipe($.concat('index.js'))
     .pipe(gulp.dest(`${distPath}/config/`))
@@ -56,7 +59,7 @@ gulp.task('default', ['clean'], function () {
   // 获取环境配置
   env = process.env.npm_config_qa ? 'qa' : process.env.npm_config_uat ? 'uat' : 'prod'
   // 开始打包编译
-  gulp.start(['replace:cdnUrl', 'replace:staticFileName', 'concat:config'], function () {
+  gulp.start(['build', 'create:versionCatalog', 'replace:cdnUrl', 'replace:version', 'concat:config'], function () {
     // 清除, 编译 / 处理项目中产生的文件
     del([`${distPath}/static`, `${versionPath}/static/config`])
   })
