@@ -1,20 +1,26 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
 import Vue from 'vue'
-import App from './App'
-import router from './router'       // api: https://router.vuejs.org/
-import store from './store'         // api: https://vuex.vuejs.org/
-import VueCookie from 'vue-cookie'  // api: https://github.com/alfhen/vue-cookie
-import '@/element-ui'               // api: https://github.com/ElemeFE/element
-import '@/iconfont'                 // api: http://www.iconfont.cn/
+import App from '@/App'
+import router from '@/router'                 // api: https://github.com/vuejs/vue-router
+import store from '@/store'                   // api: https://github.com/vuejs/vuex
+import VueCookie from 'vue-cookie'            // api: https://github.com/alfhen/vue-cookie
+import '@/element-ui'                         // api: https://github.com/ElemeFE/element
+import '@/icons'                              // api: http://www.iconfont.cn/
+import '@/element-ui-theme'
 import '@/assets/scss/index.scss'
+import httpRequest from '@/utils/httpRequest' // api: https://github.com/axios/axios
 import { isAuth } from '@/utils'
 
+Vue.use(VueCookie)
 Vue.config.productionTip = false
 
-Vue.use(VueCookie)
+// 非生产环境, 适配mockjs模拟数据                 // api: https://github.com/nuysoft/Mock
+if (process.env.NODE_ENV !== 'production') {
+  require('@/mock')
+}
 
-Vue.prototype.isAuth = isAuth // 挂载权限方法
+// 挂载全局
+Vue.prototype.$http = httpRequest // ajax请求方法
+Vue.prototype.isAuth = isAuth     // 权限方法
 
 /* eslint-disable no-new */
 new Vue({
