@@ -2,17 +2,13 @@
  * 测试环境
  */
 ;(function () {
-  window.SITE_CONFIG = {}
-  // api请求地址
-  window.SITE_CONFIG.baseUrl = '//demo.renren.io/renren-fast/'
+  window.SITE_CONFIG = {};
 
-  // 嵌套iframe地址
-  window.SITE_CONFIG.nestIframeUrl = '//demo.renren.io/renren-fast/'
-  // 嵌套iframe路由名称列表
-  window.SITE_CONFIG.nestIframeRouteNameList = ['sql']
-  
-  // 静态资源文件夹名称
-  window.SITE_CONFIG.staticFileName = ''
-  // cdn地址
-  window.SITE_CONFIG.cdnUrl = './' + window.SITE_CONFIG.staticFileName
+  // api接口请求地址
+  window.SITE_CONFIG['baseUrl'] = 'http://demo.renren.io/renren-fast';
+
+  // cdn地址 = 域名 + 版本号
+  window.SITE_CONFIG['domain']  = './'; // 域名
+  window.SITE_CONFIG['version'] = '';   // 版本号(年月日时分)
+  window.SITE_CONFIG['cdnUrl']  = window.SITE_CONFIG.domain + window.SITE_CONFIG.version;
 })();
